+++ NEWS.md
@@ -5,6 +5,10 @@
 Release notes for new releases are now on [GitHub Releases](https://github.com/electron/packager/releases)
 for the repository.
 
+### Changed
+
+* Added `osxSign.continueOnError` option. Setting it to `false` will fail the build if there was an error signing the app instead of the default behavior of printing a warning. (#1579)
+
 ## [17.1.2] - 2023-08-18
 
 [17.1.2]: https://github.com/electron/packager/compare/v17.1.1...v17.1.2